backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-pptx==1.0.2
python-docx==1.1.2
anthropic==0.40.0
python-multipart==0.0.12
requests==2.32.0
python-dotenv==1.0.1
lxml==6.0.2